Groq_API/requirements.txt
langchain
openai
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
